requirements.txt
antlr4-python3-runtime==4.9.3
docstring_parser==0.17.0
grpcio==1.76.0
hydra-core==1.3.2
Jinja2==3.1.6
jsonargparse==4.43.0
lightning==2.5.6
Markdown==3.10
MarkupSafe==2.1.5
networkx==3.5
numpy==2.3.3
omegaconf==2.3.0
packaging==25.0
pillow==11.3.0
protobuf==6.33.1
python-dotenv==1.2.1
PyYAML==6.0.3
sympy==1.14.0
tensorboard==2.20.0
torch==2.9.1+cu130
torchaudio==2.9.1
torchmetrics==1.8.2
torchvision==0.24.1+cu130
tqdm==4.67.1
triton==3.5.1